+++ src/SlimCluster.Consensus.Raft/RaftLeaderState.cs
@@ -1,0 +1,13 @@
+﻿namespace SlimCluster.Consensus.Raft;
+
+public class RaftLeaderState
+{
+    public IDictionary<string, int> NextIndex { get; protected set; }
+    public IDictionary<string, int> MatchIndex { get; protected set; }
+
+    public RaftLeaderState()
+    {
+        NextIndex = new Dictionary<string, int>();
+        MatchIndex = new Dictionary<string, int>();
+    }
+}
