+++ src/SlimCluster.Consensus.Raft/StateMachine/IStateMachine.cs
@@ -1,0 +1,8 @@
+﻿namespace SlimCluster.Consensus.Raft;
+
+public interface IStateMachine
+{
+    Task Apply(IEnumerable<object> commands);
+
+    // ToDo: Rebuild from snapshot
+}
