+++ src/SlimCluster.Consensus.Raft/Transport/AppendEntriesRequest.cs
@@ -1,0 +1,11 @@
+﻿namespace SlimCluster.Consensus.Raft;
+
+public class AppendEntriesRequest
+{
+    public int Term { get; set; }
+    public string? LeaderId { get; set; }
+    public int PrevLogIndex { get; set; }
+    public int PrevLogTerm { get; set; }
+    public IReadOnlyCollection<object>? Entries { get; set; }
+    public int LeaderCommitIndex { get; set; }
+}
