+++ src/SlimCluster.Consensus.Raft/Transport/AppendEntriesResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace SlimCluster.Consensus.Raft;
+
+public class AppendEntriesResponse
+{
+    public bool Success { get; set; }
+    public int Term { get; set; }
+}
