+++ src/SlimCluster.Consensus.Raft/Transport/InstallSnapshotRequest.cs
@@ -1,0 +1,12 @@
+﻿namespace SlimCluster.Consensus.Raft;
+
+public class InstallSnapshotRequest
+{
+    public int Term { get; set; }
+    public string? LeaderId { get; set; }
+    public int LastIncludedIndex { get; set; }
+    public int LastIncludedTerm { get; set; }
+    public int Offset { get; set; }
+    public byte[]? Data { get; set; }
+    public bool Done { get; set; }
+}
