+++ src/SlimCluster.Consensus.Raft/Transport/InstallSnapshotResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace SlimCluster.Consensus.Raft;
+
+public class InstallSnapshotResponse
+{
+    public int Term { get; set; }
+}
