+++ src/SlimCluster.Consensus.Raft/Transport/RequestVoteRequest.cs
@@ -1,0 +1,9 @@
+﻿namespace SlimCluster.Consensus.Raft;
+
+public class RequestVoteRequest
+{
+    public int Term { get; set; }
+    public string? CandidateId { get; set; }
+    public int LastLogIndex { get; set; }
+    public int LasLogTerm { get; set; }
+}
