+++ src/SlimCluster.Consensus.Raft/Transport/RequestVoteResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace SlimCluster.Consensus.Raft;
+
+public class RequestVoteResponse
+{
+    public bool VoteGranted { get; set; }
+    public int Term { get; set; }
+}
