--- src/SlimCluster.Membership.Redis/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace SlimCluster.Membership.Redis;
-
-public class Class1
-{
-}
