+++ src/SlimCluster.Membership.Swim/SlimCluster.Membership.Swim.csproj
@@ -13,7 +13,7 @@
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.1.14" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.14" />
     <PackageReference Include="Microsoft.Extensions.Options" Version="3.1.14" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
   </ItemGroup>
 
   <ItemGroup>
