--- src/SlimCluster.Strategy.Raft/SlimCluster.Strategy.Raft.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <EnableNETAnalyzers>true</EnableNETAnalyzers>
-    <LangVersion>latest</LangVersion>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SlimCluster.Membership\SlimCluster.Membership.csproj" />
-    <ProjectReference Include="..\SlimCluster.Transport\SlimCluster.Transport.csproj" />
-    <ProjectReference Include="..\SlimCluster\SlimCluster.csproj" />
-  </ItemGroup>
-
-</Project>
