--- src/SlimCluster.Strategy.Raft/StateMachine/IStateMachine.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace SlimCluster.Strategy.Raft
-{
-    public interface IStateMachine
-    {
-        Task Apply(IEnumerable<object> commands);
-
-        // ToDo: Rebuild from snapshot
-    }
-}
