--- src/SlimCluster.Strategy.Raft/Transport/AppendEntriesResponse.cs
@@ -1,8 +1,0 @@
-﻿namespace SlimCluster.Strategy.Raft
-{
-    public class AppendEntriesResponse
-    {
-        public bool Success { get; set; }
-        public int Term { get; set; }
-    }
-}
