--- src/SlimCluster.Strategy.Raft/Transport/IRaftTransport.cs
@@ -1,11 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace SlimCluster.Strategy.Raft
-{
-    public interface IRaftTransport
-    {
-        Task<RequestVoteResponse> RequestVote(RequestVoteRequest request, IAddress node);
-        Task<AppendEntriesResponse> AppendEntries(AppendEntriesRequest request, IAddress node);
-        Task<InstallSnapshotResponse> InstalSnapshot(InstallSnapshotRequest request, IAddress node);
-    }
-}
