--- src/SlimCluster.Strategy.Raft/Transport/InstallSnapshotResponse.cs
@@ -1,7 +1,0 @@
-﻿namespace SlimCluster.Strategy.Raft
-{
-    public class InstallSnapshotResponse
-    {
-        public int Term { get; set; }
-    }
-}
