--- src/SlimCluster.Strategy.Raft/Transport/RequestVoteRequest.cs
@@ -1,11 +1,0 @@
-﻿namespace SlimCluster.Strategy.Raft
-{
-    public class RequestVoteRequest
-    {
-        public int Term { get; set; }
-        public string? CandidateId { get; set; }
-        public int LastLogIndex { get; set; }
-        public int LasLogTerm { get; set; }
-    }
-
-}
