--- src/SlimCluster.Strategy.Raft/Transport/RequestVoteResponse.cs
@@ -1,9 +1,0 @@
-﻿namespace SlimCluster.Strategy.Raft
-{
-    public class RequestVoteResponse
-    {
-        public bool VoteGranted { get; set; }
-        public int Term { get; set; }
-    }
-
-}
