--- src/SlimCluster.Transport/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace SlimCluster.Transport;
-
-public class Class1
-{
-}
