+++ src/SlimCluster.sln
@@ -5,11 +5,7 @@ VisualStudioVersion = 17.0.32126.317
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SlimCluster", "SlimCluster\SlimCluster.csproj", "{77A1D853-5C45-4DAD-B2F0-0B4F6BB55B2F}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SlimCluster.MsDependencyInjection", "SlimCluster.MsDependencyInjection\SlimCluster.MsDependencyInjection.csproj", "{50B8ED81-C172-4302-B327-AF741B05B9F4}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SlimCluster.Strategy.Raft", "SlimCluster.Strategy.Raft\SlimCluster.Strategy.Raft.csproj", "{7199B34A-3DAD-43EC-990A-9F5C89FA6F0A}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SlimCluster.Transport", "SlimCluster.Transport\SlimCluster.Transport.csproj", "{05F67E48-1F66-48F0-A876-4A6C55C669BC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SlimCluster.Consensus.Raft", "SlimCluster.Consensus.Raft\SlimCluster.Consensus.Raft.csproj", "{7199B34A-3DAD-43EC-990A-9F5C89FA6F0A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{525A6FC4-A561-4A56-8830-B5EB9A132EB4}"
 	ProjectSection(SolutionItems) = preProject
@@ -21,8 +17,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{42420C23-31F6-46F4-AFEA-A12206C47366}"
 	ProjectSection(SolutionItems) = preProject
-		Samples\Dockerfile = Samples\Dockerfile
 		Samples\Docker-Build.ps1 = Samples\Docker-Build.ps1
+		Samples\Dockerfile = Samples\Dockerfile
 		Samples\README.md = Samples\README.md
 	EndProjectSection
 EndProject
@@ -50,18 +46,10 @@ Global
 		{77A1D853-5C45-4DAD-B2F0-0B4F6BB55B2F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{77A1D853-5C45-4DAD-B2F0-0B4F6BB55B2F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{77A1D853-5C45-4DAD-B2F0-0B4F6BB55B2F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{50B8ED81-C172-4302-B327-AF741B05B9F4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{50B8ED81-C172-4302-B327-AF741B05B9F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{50B8ED81-C172-4302-B327-AF741B05B9F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{50B8ED81-C172-4302-B327-AF741B05B9F4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7199B34A-3DAD-43EC-990A-9F5C89FA6F0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7199B34A-3DAD-43EC-990A-9F5C89FA6F0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7199B34A-3DAD-43EC-990A-9F5C89FA6F0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7199B34A-3DAD-43EC-990A-9F5C89FA6F0A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{05F67E48-1F66-48F0-A876-4A6C55C669BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{05F67E48-1F66-48F0-A876-4A6C55C669BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{05F67E48-1F66-48F0-A876-4A6C55C669BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{05F67E48-1F66-48F0-A876-4A6C55C669BC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AC8D5966-70AC-44A6-BB8C-4946B2D5EA9F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AC8D5966-70AC-44A6-BB8C-4946B2D5EA9F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AC8D5966-70AC-44A6-BB8C-4946B2D5EA9F}.Release|Any CPU.ActiveCfg = Release|Any CPU
